Run ratchet Docker container as the default user
This is a trade-off between security and usability with GitHub Actions. Since this is a static binary in a scratch container that is not a long-running service, the trade-off of running as root (so it works inside GitHub Actions workflows) is probably worth the risk of running the CLI as the default Docker user.